Breaking Changes feat: added more placeholders removed: removed screens folder removed: removed home component changes: using expo router changes: using bottomtabnvaigator from react changes: renamed components to be more consise changes: index.tsx is now homepage
