+++ project1/app/(tabs)/settings.tsx
@@ -1,0 +1,1 @@
+// placeholder for a settings page need to be worked on
